stage-2/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
pymysql==1.1.1
httpx==0.27.2
python-dotenv==1.0.1
Pillow==10.4.0
pydantic==2.9.2
